inventory controls responding in input controller
